chatbot/requirements.txt
PyPDF2>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=0.27.0
requests>=2.28.0
